pyproject.toml
[tool.poetry]
authors = ["Robert Kaussow <mail@thegeeklab.de>"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Natural Language :: English",
  "Operating System :: POSIX",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Utilities",
  "Topic :: Software Development",
]
description = "Clone single branch from all repositories listed in a file."
documentation = "https://github.com/thegeeklab/git-batch/"
homepage = "https://github.com/thegeeklab/git-batch/"
include = ["LICENSE"]
keywords = ["git", "batch", "automation"]
license = "MIT"
name = "git-batch"
packages = [{ include = "gitbatch", format = ["sdist", "wheel"] }]
readme = "README.md"
repository = "https://github.com/thegeeklab/git-batch/"
version = "0.0.0"

[tool.poetry.dependencies]
GitPython = "3.1.45"
colorama = "0.4.6"
python = "^3.11.0"
python-json-logger = "4.0.0"
pylint = "^4.0.3"

[tool.poetry.scripts]
git-batch = "gitbatch.cli:main"

[tool.poetry.group.dev.dependencies]
ruff = "0.14.5"
pytest = "9.0.1"
pytest-mock = "3.15.1"
pytest-cov = "7.0.0"
toml = "0.10.2"
mypy = "1.10.0"

[tool.poetry-dynamic-versioning]
enable = true
style = "semver"
vcs = "git"

[tool.pytest.ini_options]
addopts = "gitbatch --cov=gitbatch --cov-report=xml:coverage.xml --cov-report=term-missing --no-cov-on-fail"
filterwarnings = [
  "ignore::FutureWarning",
  "ignore::DeprecationWarning",
  "ignore:.*pep8.*:FutureWarning",
]

[tool.coverage.run]
omit = ["**/test/*"]

[build-system]
build-backend = "poetry_dynamic_versioning.backend"
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]

[tool.ruff]
exclude = [
  ".git",
  "__pycache__",
  "build",
  "dist",
  "test",
  "*.pyc",
  "*.egg-info",
  ".cache",
  ".eggs",
  "env*",
]

line-length = 99
indent-width = 4

[tool.ruff.lint]
# Explanation of errors
#
# D102: Missing docstring in public method
# D103: Missing docstring in public function
# D105: Missing docstring in magic method
# D107: Missing docstring in __init__
# D202: No blank lines allowed after function docstring
# D203: One blank line required before class docstring
# D212: Multi-line docstring summary should start at the first line
ignore = ["D102", "D103", "D105", "D107", "D202", "D203", "D212", "RUF012"]
select = [
  "D",
  "E",
  "F",
  "Q",
  "W",
  "I",
  "S",
  "BLE",
  "N",
  "UP",
  "B",
  "A",
  "C4",
  "T20",
  "SIM",
  "RET",
  "ARG",
  "ERA",
  "RUF",
]

[tool.ruff.lint.flake8-builtins]
builtins-allowed-modules = ["logging", "copy"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
explicit_package_bases = true
no_implicit_optional = false
strict_optional = false
allow_redefinition = true
allow_untyped_globals = true
local_partial_types = false
warn_unused_ignores = true
check_untyped_defs = true
ignore_missing_imports = true
files = ["gitbatch/"]
